Restructuring of project to align with gargoyle s3 proxy project
- Proper use of reference.conf vs application.conf
- Setup integration tests with fixtures and start on random port
- package structure
- use settings loaded from actor system
- update build.sbt
