+++ src/main/resources/application.conf
@@ -5,17 +5,7 @@ akka {
   actor {
     debug.unhandled = on
   }
-}
-
-akka.http.server {
-  verbose-error-messages = on
-}
-
-api.server {
-  port = 12345
-  port = ${?STS_PORT}
-  interface = 0.0.0.0
-  interface = ${?STS_INTERFACE}
-  hostname = localhost
-  hostname = ${?STS_HOSTNAME}
+  http.server {
+    verbose-error-messages = on
+  }
 }
