+++ src/main/resources/reference.conf
@@ -1,0 +1,9 @@
+gargoyle {
+  sts {
+    # Default interface and port to listen on.
+    http {
+      bind = "127.0.0.1"
+      port = 12345
+    }
+  }
+}
