+++ src/main/scala/com/ing/wbaa/gargoyle/sts/Server.scala
@@ -1,0 +1,7 @@
+package com.ing.wbaa.gargoyle.sts
+
+import akka.actor.ActorSystem
+
+object Server extends App {
+  val stsService: StsService = StsService()(ActorSystem.create("gargoyle-sts"))
+}
