--- src/main/scala/ing/wbaa/gargoyle/sts/CoreActorSystem.scala
@@ -1,19 +1,0 @@
-package ing.wbaa.gargoyle.sts
-
-import akka.actor.ActorSystem
-import akka.stream.ActorMaterializer
-import com.typesafe.config.{ Config, ConfigFactory }
-
-trait CoreActorSystem {
-  implicit def system: ActorSystem = ActorSystem()
-
-  sys.addShutdownHook {
-    system.terminate()
-  }
-  val config: Config = ConfigFactory.load()
-}
-
-trait Actors {
-  this: CoreActorSystem =>
-  implicit val materializer: ActorMaterializer = ActorMaterializer()
-}
