--- src/main/scala/ing/wbaa/gargoyle/sts/Server.scala
@@ -1,8 +1,0 @@
-package ing.wbaa.gargoyle.sts
-
-object Server extends App
-  with CoreActorSystem
-  with Routes
-  with Actors
-  with Web {
-}
